+++ Pipfile
@@ -20,6 +20,7 @@ typing-extensions = "*"
 flask-jwt-extended = "==4.6.0"
 wtforms = "==3.1.2"
 sqlalchemy = "*"
+bcrypt =  "*"
 
 [requires]
 python_version = "3.13"
