+++ src/api/utils.py
@@ -1,5 +1,8 @@
+import os
 from flask import jsonify, url_for
 from itsdangerous import URLSafeTimedSerializer
+import smtplib
+from email.mime.text import MIMEText
 
 class APIException(Exception):
     status_code = 400
@@ -64,3 +67,35 @@ class SerializerSingleton:
 
     def loads(self, token):
         return self.serializer.loads(token, salt="password-reset", max_age=300)
+    
+def send_email(to, url):
+    SMTP_SERVER = "in-v3.mailjet.com"
+    SMTP_PORT = 587
+    EMAIL_FROM = os.getenv("EMAIL_USER")
+    SMTP_USER = os.getenv("SMTP_USER")
+    SMTP_PASSWORD = os.getenv("SMTP_PASSWORD")
+    if not all([EMAIL_FROM, SMTP_USER, SMTP_PASSWORD]):
+        print("Error: Email credentials not configured")
+        return False
+
+    message = f"""
+        <p>¡Hi!</p>
+        <p>You have requested to reset your password
+        Click <a href="{url}">here</a> to continue</p>
+        <p><i>If you didn't request this, please ignore this email</i></p>
+        """
+
+    msg = MIMEText(message, "html")
+    msg["Subject"] = "Password reset"
+    msg["From"] = EMAIL_FROM
+    msg["To"] = to
+
+    try:
+        with smtplib.SMTP(SMTP_SERVER, SMTP_PORT) as server:
+            server.starttls()
+            server.login(SMTP_USER, SMTP_PASSWORD)
+            server.sendmail(EMAIL_FROM, [to], msg.as_string())
+        return True
+    except Exception as e:
+        print(f"Error sending email: {e}")
+        return False
